Move impl to correct place
