+++ imessage-database/src/tables/chat_handle.rs
@@ -4,7 +4,7 @@
 
 use std::collections::{BTreeSet, HashMap};
 
-use crate::tables::table::{Cacheable, Table, CHAT_HANDLE_JOIN};
+use crate::tables::table::{Cacheable, Deduplicate, Table, CHAT_HANDLE_JOIN};
 use rusqlite::{Connection, Error, Result, Row, Statement};
 
 /// Represents a single row in the `chat_handle_join` table.
@@ -79,6 +79,36 @@ impl Cacheable for ChatToHandle {
 
         cache
     }
-
-    // TODO: Implement Diagnostic, determine how many chats do not exist in the join table
 }
+
+impl Deduplicate for ChatToHandle {
+    type T = BTreeSet<i32>;
+
+    /// Given the initial set of duplciated chats, deduplciate them based on the participants
+    ///
+    /// This returns a new hashmap that maps the real chat ID to a new deduplicated unique chat ID
+    /// that represents a single chat for all of the same participants, even if they have multiple handles
+    fn dedupe(duplicated_data: &HashMap<i32, Self::T>) -> HashMap<i32, i32> {
+        let mut deduplicated_chats: HashMap<i32, i32> = HashMap::new();
+        let mut participants_to_unique_chat_id: HashMap<Self::T, i32> = HashMap::new();
+
+        // Build cache of each unique set of participants to a new identifier:
+        let mut unique_chat_identifier = 0;
+        for (chat_id, participants) in duplicated_data {
+            match participants_to_unique_chat_id.get(participants) {
+                Some(id) => {
+                    deduplicated_chats.insert(chat_id.to_owned(), id.to_owned());
+                }
+                None => {
+                    participants_to_unique_chat_id
+                        .insert(participants.to_owned(), unique_chat_identifier);
+                    deduplicated_chats.insert(chat_id.to_owned(), unique_chat_identifier);
+                    unique_chat_identifier += 1;
+                }
+            }
+        }
+        deduplicated_chats
+    }
+}
+
+// TODO: Implement Diagnostic, determine how many chats do not exist in the join table
